Disable output on createIndex
I followed your tutorial here http://tnt.studio/blog/did-you-mean-functionality-with-laravel-scout where you explain how to create trigrams.
If the function `createIndex` is called there is no way to disable the ouput of the `TNTINdexer`.
This is suboptimal if you have an api whereof you can create posts or sth. else and run the trigram generator directly after the creation process.